Crate aws_smithy_http

Core HTTP primitives for service clients generated by smithy-rs including:

HTTP Body implementation

Endpoint support

HTTP header deserialization

Event streams

ByteStream: a misuse-resistant abstraction for streaming binary data

Feature	Description

rt-tokio	Provides features that are dependent on tokio including the ByteStream::from_path util

event-stream	Provides Sender/Receiver implementations for Event Stream codegen.

Modules

body

byte_stream

ByteStream Abstractions

callback

A module for traits that define callbacks that will be called at specific points in an HTTP request’s lifecycle.

endpoint

event_stream

Provides Sender/Receiver implementations for Event Stream codegen.

header

Utilities for parsing information from headers

http_versions

HTTP Version-related code

label

Formatting values as Smithy httpLabel

middleware

This modules defines the core, framework agnostic, HTTP middleware interface used by the SDK

operation

property_bag

query

Utilities for writing Smithy values into a query string.

response

result

Result wrapper types for success and failure responses.

retry

HTTP specific retry behaviors

Macros

pin_mut

Pins a value on the stack.