Struct aws_smithy_http::body::SdkBody

pub struct SdkBody { /* private fields */ }

SdkBody type

This is the Body used for dispatching all HTTP Requests. For handling responses, the type of the body will be controlled by the HTTP stack.

TODO(naming): Consider renaming to simply Body, although I’m concerned about naming headaches between hyper::Body and our Body

Implementations

impl SdkBody

pub fn from_dyn(body: BoxBody<Bytes, Error>) -> Self

Construct an SdkBody from a Boxed implementation of http::Body

pub fn retryable(f: impl Fn() -> SdkBody + Send + Sync + 'static) -> Self

Construct an explicitly retryable SDK body

Note: This is probably not what you want

All bodies constructed from in-memory data (String, Vec<u8>, Bytes, etc.) will be retryable out of the box. If you want to read data from a file, you should use ByteStream::from_path. This function is only necessary when you need to enable retries for your own streaming container.

pub fn taken() -> Self

pub fn empty() -> Self

pub fn bytes(&self) -> Option<&[u8]>

If possible, return a reference to this body as &[u8]

If this SdkBody is NOT streaming, this will return the byte slab If this SdkBody is streaming, this will return None

pub fn try_clone(&self) -> Option<Self>

pub fn content_length(&self) -> Option<u64>

pub fn with_callback(&mut self, callback: Box<dyn BodyCallback>) -> &mut Self