aws_smithy_http

Module event_stream

Provides Sender/Receiver implementations for Event Stream codegen.

Structs

EventStreamSender

Input type for Event Streams.

MessageStreamAdapter

Adapts a Stream<SmithyMessageType> to a signed Stream<Bytes> by using the provided message marshaller and signer implementations.

MessageStreamError

An error that occurs within a message stream.

Receiver

Receives Smithy-modeled messages out of an Event Stream.

ReceiverError

An error that occurs within an event stream receiver.

Enums

InitialMessageType

Type Aliases

BoxError

A generic, boxed error that’s Send, Sync, and 'static.