Module aws_smithy_http::header

Utilities for parsing information from headers

Structs

ParseError

Functions

headers_for_prefix

many_dates

Read all the dates from the header map at key according the format

one_or_none

Read exactly one or none from a headers iterator

quote_header_value

Conditionally quotes and escapes a header value if the header value contains a comma or quote.

read_many_from_str

read_many_primitive

set_request_header_if_absent

set_response_header_if_absent