aws_smithy_http::header

Function append_merge_header_maps

pub fn append_merge_header_maps(
    lhs: HeaderMap<HeaderValue>,
    rhs: HeaderMap<HeaderValue>,
) -> HeaderMap<HeaderValue>

Given two HeaderMaps, merge them together and return the merged HeaderMap. If the two HeaderMaps share any keys, values from the right HeaderMap be appended to the left HeaderMap.