aws_smithy_http::header

Function append_merge_header_maps_http_1x

pub fn append_merge_header_maps_http_1x(
    lhs: HeaderMap<HeaderValue>,
    rhs: HeaderMap<HeaderValue>,
) -> HeaderMap<HeaderValue>

Given two http-1x HeaderMaps, merge them together and return the merged HeaderMap. If the two HeaderMaps share any keys, values from the right HeaderMap be appended to the left HeaderMap.