Function aws_smithy_http::header::headers_for_prefix

pub fn headers_for_prefix<'a>(
    header_names: impl Iterator<Item = &'a str>,
    key: &'a str,
) -> impl Iterator<Item = (&'a str, &'a str)>

Returns an iterator over pairs where the first element is the unprefixed header name that starts with the input key prefix, and the second element is the full header name.