Function aws_smithy_http::middleware::load_response

pub async fn load_response<T, E, O>(
    response: Response,
    handler: &O
) -> Result<SdkSuccess<T>, SdkError<E>>where
    O: ParseHttpResponse<Output = Result<T, E>>,

Load a response using handler to parse the results.

This function is intended to be used on the response side of a middleware chain.

Success and failure will be split and mapped into SdkSuccess and SdkError. Generic Parameters:

O: The Http response handler that returns Result<T, E>

T/E: Result<T, E> returned by handler.