Crate aws_smithy_http

Core HTTP primitives for service clients generated by smithy-rs including:

HTTP Body implementation

Endpoint support

HTTP header deserialization

Event streams

Feature	Description

rt-tokio	Provides features that are dependent on tokio including the ByteStream::from_path util

event-stream	Provides Sender/Receiver implementations for Event Stream codegen.

Modules

endpoint

Code for resolving an endpoint (URI) that a request should be sent to

header

Utilities for parsing information from headers

label

Formatting values as Smithy httpLabel

operation

Deprecated metadata type.

query

Utilities for writing Smithy values into a query string.